Solution to Logical AND and OR Combination

+++ Logical AND and OR Combination@@67882a731b4392d8fc1577df/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main(){
-    float a,b;
-    scanf("%f %f",&a,&b);
+    int a,b;
+    scanf("%d %d",&a,&b);
     if(a>0 && b>0||a==0 && b==0){
         printf("True");
     }
